+++ src/main/java/com/boot/rest/Application.java
@@ -5,6 +5,7 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.security.SecurityAutoConfiguration;
 import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.ComponentScan;
 import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
 import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseBuilder;
 import org.springframework.jdbc.datasource.embedded.EmbeddedDatabaseType;
@@ -30,7 +31,8 @@ import java.util.Properties;
 @EnableTransactionManagement
 @EnableJpaRepositories("com.boot.rest.dao")
 @EnableWebSecurity
-public class Application extends WebSecurityConfigurerAdapter {
+@ComponentScan("com.boot.rest")
+public class Application {
 
     @Bean
     public DataSource dataSource() {
@@ -68,30 +70,30 @@ public class Application extends WebSecurityConfigurerAdapter {
         return jpaTransactionManager;
     }
 
-    @Override
-    public void configure(HttpSecurity http) throws Exception {
-        http
-                .csrf().disable()
-                .authorizeRequests()
-                .antMatchers("/").hasRole("USER")
-                .antMatchers("/index.html").hasRole("USER")
-                .antMatchers("/admin.html").hasRole("ADMIN")
-                .antMatchers("/data/**/").hasRole("API")
-                .and()
-                .authorizeRequests().anyRequest().fullyAuthenticated()
-                .and()
-                .httpBasic()
-                .and()
-                .formLogin();
-    }
+//    @Override
+//    public void configure(HttpSecurity http) throws Exception {
+//        http
+//                .csrf().disable()
+//                .authorizeRequests()
+//                .antMatchers("/").hasRole("USER")
+//                .antMatchers("/index.html").hasRole("USER")
+//                .antMatchers("/admin.html").hasRole("ADMIN")
+//                .antMatchers("/data/**/").hasRole("API")
+//                .and()
+//                .authorizeRequests().anyRequest().fullyAuthenticated()
+//                .and()
+//                .httpBasic()
+//                .and()
+//                .formLogin();
+//    }
 
-    @Autowired
-    public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
-        auth.inMemoryAuthentication()
-                .withUser("user").password("pass").roles("USER", "API").and()
-                .withUser("admin").password("pass").roles("ADMIN", "USER", "API").and()
-                .withUser("noapi").password("pass").roles("USER");
-    }
+//    @Autowired
+//    public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
+//        auth.inMemoryAuthentication()
+//                .withUser("user").password("pass").roles("USER", "API").and()
+//                .withUser("admin").password("pass").roles("ADMIN", "USER", "API").and()
+//                .withUser("noapi").password("pass").roles("USER");
+//    }
 
     public static void main(String[] args) throws Throwable {
         SpringApplication.run(Application.class, args);
