+++ src/main/java/com/boot/rest/dao/UserRepository.java
@@ -10,5 +10,5 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserRepository extends JpaRepository<User, Integer> {
-    User findByUserName(String username);
+    User findByUsername(String username);
 }
