+++ src/main/java/com/boot/rest/service/MyUserDetailsService.java
@@ -1,18 +1,17 @@
-package com.boot.rest.config;
+package com.boot.rest.service;
 
 import com.boot.rest.dao.UserRepository;
 import com.boot.rest.model.User;
 import com.boot.rest.model.UserRole;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.config.annotation.authentication.configurers.GlobalAuthenticationConfigurerAdapter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
-import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
 import java.util.stream.Collectors;
@@ -21,6 +20,7 @@ import java.util.stream.Collectors;
  * Created by Oleg on 26 Jul 2016.
  */
 
+@Service("userDetailsService")
 public class MyUserDetailsService implements UserDetailsService {
 
     @Autowired
@@ -30,7 +30,7 @@ public class MyUserDetailsService implements UserDetailsService {
     public UserDetails loadUserByUsername(final String username)
             throws UsernameNotFoundException {
 
-        User user = userRepository.findByUserName(username);
+        User user = userRepository.findByUsername(username);
         List<GrantedAuthority> authorities = buildUserAuthority(user.getUserRole());
 
         return buildUserForAuthentication(user, authorities);
